refactor(nvim): migrate from AstroNvim to LazyVim
