--- dot_config/lazyvim/init.lua
@@ -1,2 +1,0 @@
--- bootstrap lazy.nvim, LazyVim and your plugins
-require("config.lazy")
