--- dot_config/lazyvim/lua/plugins/snacks.lua
@@ -1,25 +1,0 @@
-return {
-  "snacks.nvim",
-  ---@type snacks.Config
-  opts = {
-    indent = {
-      enabled = true,
-      chunk = {
-        enabled = true,
-        char = {
-          -- corner_top = "┌",
-          -- corner_bottom = "└",
-          corner_top = "╭",
-          corner_bottom = "╰",
-          horizontal = "─",
-          vertical = "│",
-          arrow = ">",
-        },
-      },
-      animate = {
-        enabled = true,
-        style = "down",
-      },
-    },
-  },
-}
