--- dot_config/nvim/.luarc.json
@@ -1,15 +1,0 @@
-{
-  "format": {
-    "enable": false
-  },
-  "runtime": {
-    "version": "LuaJIT"
-  },
-  "diagnostics": {
-    "globals": ["vim", "Snacks", "LazyVim"],
-    "disable": ["mixed_table", "undefined-doc-name", "missing-parameter", "param-type-mismatch"]
-  },
-  "workspace": {
-    "checkThirdParty": false
-  }
-}
