--- dot_config/nvim/after/ftplugin/python.lua
@@ -1,5 +1,0 @@
-local wk = require("which-key")
-
-wk.add({
-  { "<leader>lv", "<cmd>VenvSelect<CR>", desc = "Select venv" },
-})
