+++ dot_config/nvim/lua/config/options.lua
