--- dot_config/nvim/lua/plugins/flash.lua
@@ -1,11 +1,0 @@
-return {
-  "flash.nvim",
-  event = "VeryLazy",
-  opts = {
-    modes = {
-      search = {
-        enabled = true,
-      },
-    },
-  },
-}
