--- dot_config/nvim/lua/plugins/mason.lua
@@ -1,33 +1,0 @@
--- Customize Mason
-
----@type LazySpec
-return {
-  -- use mason-tool-installer for automatically installing Mason packages
-  {
-    "WhoIsSethDaniel/mason-tool-installer.nvim",
-    -- overrides `require("mason-tool-installer").setup(...)`
-    opts = {
-      -- Make sure to use the names found in `:Mason`
-      ensure_installed = {
-        -- install language servers
-        "lua-language-server",
-        "rust-analyzer",
-
-        -- install linters
-        "luacheck",
-        "yamllint",
-
-        -- install formatters
-        "stylua",
-        "prettier",
-        "yamlfmt",
-
-        -- install debuggers
-        "debugpy",
-
-        -- install any other package
-        "tree-sitter-cli",
-      },
-    },
-  },
-}
