--- dot_config/nvim/lua/plugins/smear-cursor.lua
@@ -1,5 +1,0 @@
-return {
-  "sphamba/smear-cursor.nvim",
-  event = "VeryLazy",
-  opts = {},
-}
