+++ dot_config/nvim/lua/plugins/snacks.lua
@@ -1,40 +1,25 @@
 return {
   "snacks.nvim",
-  event = "VeryLazy",
+  ---@type snacks.Config
   opts = {
-    dashboard = {
-      preset = {
-        pick = function(cmd, opts) return LazyVim.pick(cmd, opts)() end,
-        header = table.concat({
-          " █████  ███████ ████████ ██████   ██████ ",
-          "██   ██ ██         ██    ██   ██ ██    ██",
-          "███████ ███████    ██    ██████  ██    ██",
-          "██   ██      ██    ██    ██   ██ ██    ██",
-          "██   ██ ███████    ██    ██   ██  ██████ ",
-          "",
-          "███    ██ ██    ██ ██ ███    ███",
-          "████   ██ ██    ██ ██ ████  ████",
-          "██ ██  ██ ██    ██ ██ ██ ████ ██",
-          "██  ██ ██  ██  ██  ██ ██  ██  ██",
-          "██   ████   ████   ██ ██      ██",
-        }, "\n"),
-       -- stylua: ignore
-       ---@type snacks.dashboard.Item[]
-       keys = {
-         { icon = " ", key = "f", desc = "Find File", action = ":lua Snacks.dashboard.pick('files')" },
-         { icon = " ", key = "n", desc = "New File", action = ":ene | startinsert" },
-         { icon = "", key = "p", desc = "Projects", action = ":lua Snacks.dashboard.pick('projects')" },
-         { icon = " ", key = "g", desc = "Find Text", action = ":lua Snacks.dashboard.pick('live_grep')" },
-         { icon = " ", key = "r", desc = "Recent Files", action = ":lua Snacks.dashboard.pick('oldfiles')" },
-         -- { icon = " ", key = "c", desc = "Config", action = ":lua Snacks.dashboard.pick('files', {cwd = vim.fn.stdpath('config')})" },
-         { icon = " ", key = "s", desc = "Restore Session", section = "session" },
-         { icon = "󰒲 ", key = "l", desc = "Lazy", action = ":Lazy" },
-         { icon = " ", key = "q", desc = "Quit", action = ":qa" },
-       },
-      },
-    },
-    scroll = {
+    indent = {
       enabled = true,
+      chunk = {
+        enabled = true,
+        char = {
+          -- corner_top = "┌",
+          -- corner_bottom = "└",
+          corner_top = "╭",
+          corner_bottom = "╰",
+          horizontal = "─",
+          vertical = "│",
+          arrow = ">",
+        },
+      },
+      animate = {
+        enabled = true,
+        style = "down",
+      },
     },
   },
 }
