+++ dot_config/nvim/lua/plugins/tiny-inline-diagnostic.lua
