--- dot_config/nvim/lua/plugins/treesitter.lua
@@ -1,13 +1,0 @@
--- Customize Treesitter
-
----@type LazySpec
-return {
-  "nvim-treesitter/nvim-treesitter",
-  opts = {
-    ensure_installed = {
-      "lua",
-      "vim",
-      -- add more arguments for adding more treesitter parsers
-    },
-  },
-}
