--- dot_config/nvim/neovim.yml
@@ -1,10 +1,0 @@
----
-base: lua51
-
-globals:
-  vim:
-    any: true
-  Snacks:
-    any: true
-  LazyVim:
-    any: true
