--- dot_config/nvim/selene.toml
@@ -1,8 +1,0 @@
-std = "neovim"
-
-[rules]
-global_usage = "allow"
-if_same_then_else = "allow"
-incorrect_standard_library_use = "allow"
-mixed_table = "allow"
-multiple_statements = "allow"
